Specialize previous OAuth token `URLRequest` logic for Google
